Moves PintJSONEncoder to seed/serializers
We'll need this more generally, plus some other special-treatment for
Quantity objects headed out via the API. Adds a module to give these a
place to grow.
